+++ CHANGELOG.md
@@ -1,6 +1,18 @@
+## 4.0.3
+  - Update gemspec summary
+
+## 4.0.2
+  - Tighten gem deps for the 'redis' gem
+## 4.0.1
+  - Fix some documentation issues
+
+## 4.0.0
+  - Mark deprecated `name` and `queue` configuration parameters as obsolete
+
+## 3.0.3
+  - Fix logging method signature for #debug
 ## 3.0.2
   - Relax constraint on logstash-core-plugin-api to >= 1.60 <= 2.99
-
 ## 3.0.1
   - Republish all the gems under jruby.
 ## 3.0.0